requirements.txt
autogen>=0.9.7
beautifulsoup4>=4.14.3
boto3>=1.42.39
google-genai>=1.47.0
ipdb>=0.13.13
matplotlib>=3.9.4
numpy<2.0.0
pandas>=2.3.3
python-dotenv>=1.0.1
requests>=2.32.5
scikit-learn>=1.6.1
scipy>=1.13.1
seaborn>=0.13.2
termcolor>=3.1.0
tqdm>=4.67.2
trace-opt>=0.1.3.8
